Characters: hide hidden intimacies in API
Currently, hidden intimacies are fully exposed through the API and
hidden in the frontend Javascript. This change modifies the API to only
return hidden intimacies for characters where the player should be able
to see them, using the same heuristic as the frontend uses currently -
update permission.

This hides player secrets from prying eyes - darn Night castes always
on the lookout for an advantage!